NetFx Project upgrade and move common sources to shared props
