+++ src/Directory.Build.props
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Latest" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  
+  <!-- Possible values for "TargetGroup": netfx, netcoreapp, netstandard -->
+  <PropertyGroup>
+    <TargetGroup Condition="$(TargetFramework.StartsWith('netcoreapp'))">netcoreapp</TargetGroup>
+    <TargetGroup Condition="$(TargetFramework.StartsWith('netstandard'))">netstandard</TargetGroup>
+    <TargetGroup Condition="'$(TargetGroup)' != 'netcoreapp' and '$(TargetGroup)' != 'netstandard'">netfx</TargetGroup>
+  </PropertyGroup>
   <PropertyGroup>
     <TargetsWindows Condition="'$(OS)' == 'Windows_NT' AND '$(OSGroup)' == ''">true</TargetsWindows>
     <TargetsWindows Condition="'$(OS)' != 'Windows_NT' AND '$(OSGroup)' == ''">false</TargetsWindows>
@@ -20,6 +25,7 @@
     <ManagedSourceCode>$(ProjectDir)Microsoft.Data.SqlClient\</ManagedSourceCode>
     <NetCoreSource>$(ManagedSourceCode)netcore\</NetCoreSource>
     <NetFxSource>$(ManagedSourceCode)netfx\</NetFxSource>
+    <CommonSource>$(ManagedSourceCode)src\</CommonSource>
     <NetFxResources>$(ManagedSourceCode)netfx\src\Resources\</NetFxResources>
     <ObjFolder>$(RepoRoot)obj\</ObjFolder>
     <CommonPath>$(NetCoreSource)src\Common\src</CommonPath>
